% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cal_data}
\alias{cal_data}
\title{Merged PurpleAir and reference data for calibration examples}
\format{
\subsection{\code{cal_data}}{

A data frame with 1,768 rows and 24 columns:
\describe{
\item{datetime}{Hourly date-time}
\item{ref_PM25}{Reference PM2.5 (ug/m^3)}
\item{p_0_3_um_ave, p_0_5_um_ave}{Particle count numbers less than specific diameters}
...
}
}
}
\source{
Martindale et al 2025 (DOI to come)
}
\usage{
cal_data
}
\description{
A subset of data from Martindale et al 2025 to be used in examples
}
\keyword{datasets}
