% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_ALT_method}
\alias{calibrate_ALT_method}
\title{Use Lance Wallace's ALT method to correct PurpleAir measurements (with a
default calibration factor of 3).}
\usage{
calibrate_ALT_method(data, cf = NULL)
}
\arguments{
\item{data}{cleaned PurpleAir data with averaged count channels}

\item{cf}{calibration factor. This defaults to 3 (see https://www.mdpi.com/1424-8220/22/7/2755).}
}
\value{
PM2.5 measurements corrected with Lance Wallace's ALT method
(https://doi.org/10.1016/j.atmosenv.2021.118432)
}
\description{
Use Lance Wallace's ALT method to correct PurpleAir measurements (with a
default calibration factor of 3).
}
\examples{
# This assumes that the data has already been inspected.
ALT_PA_data <- calibrate_ALT_method(cal_data)
ALT_PA_data <- calibrate_ALT_method(cal_data, 3.4)
}
