% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv_data}
\alias{cv_data}
\title{Cross-validated merged data}
\format{
\subsection{\code{cv_data}}{

A dataframe of 4,704 rows and 40 columns.
\describe{
\item{datetime}{Date time}
\item{ref_PM25}{Reference PM2.5}
\item{station}{Reference station}
...
}
}
}
\source{
Martindale et al 2025 (DOI to come)
}
\usage{
cv_data
}
\description{
A subset of data from Martindale et al 2025 to be used in examples.
This is not the selected model from the paper -- it is only to be used in examples.
}
\keyword{datasets}
