% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cv.R
\name{do_CV_mult}
\alias{do_CV_mult}
\title{Do cross-validation with multiple formulas.}
\usage{
do_CV_mult(data, id, group, formulas)
}
\arguments{
\item{data}{dataframe containing the training data.}

\item{id}{unique variable for determining sort order of data frame}

\item{group}{grouping variable (a variable in the data frame)}

\item{formulas}{list of formulas to pass to linear models.}
}
\value{
dataframe with CV_preds and AIC columns for each variable
}
\description{
Do cross-validation with multiple formulas.
}
\examples{
formulas <- list(formula1 = ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity + I(current_humidity^2) +
   current_temp_f + as.factor(season),
                 formula2 = ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity + current_temp_f + as.factor(season),
                 formula3 = ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_temp_f + as.factor(season),
                 formula4 = ref_PM25 ~ 0 + pm2_5_cf_ave + as.factor(season))
train_data <- train_test$train
cv_data <- do_CV_mult(train_data, "train_index", "select_index", formulas)
}
