% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cv.R
\name{get_MSE}
\alias{get_MSE}
\title{Get mean-squared error (MSE)}
\usage{
get_MSE(obs, pred, AIC_var)
}
\arguments{
\item{obs}{outcome variable}

\item{pred}{prediction variable from a model}

\item{AIC_var}{AIC variable}
}
\value{
dataframe of RMSE, MSE-based R^2, and AIC.
}
\description{
Get mean-squared error (MSE)
}
\examples{
formulas <- list(formula1 = ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity + I(current_humidity^2) +
   current_temp_f + as.factor(season),
                 formula2 = ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity + current_temp_f + as.factor(season),
                 formula3 = ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_temp_f + as.factor(season),
                 formula4 = ref_PM25 ~ 0 + pm2_5_cf_ave + as.factor(season))
train_data <- train_test$train
train_data_CV <- do_CV_mult(train_data, "train_index", "select_index", formulas)
one_MSE <- get_MSE(train_data_CV$ref_PM25, train_data_CV$cvpreds_model_1, train_data_CV$AIC_model_1)
rbind(get_MSE(train_data_CV$ref_PM25, train_data_CV$cvpreds_model_1, train_data_CV$AIC_model_1),
  get_MSE(train_data_CV$ref_PM25, train_data_CV$cvpreds_model_2, train_data_CV$AIC_model_2),
  get_MSE(train_data_CV$ref_PM25, train_data_CV$cvpreds_model_3, train_data_CV$AIC_model_3),
  get_MSE(train_data_CV$ref_PM25, train_data_CV$cvpreds_model_4, train_data_CV$AIC_model_4))
}
