% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcscal-package.R
\docType{package}
\name{lcscal-package}
\alias{lcscal}
\alias{lcscal-package}
\title{lcscal: Calibrating Low-Cost Sensors Against Regulatory Monitors}
\description{
This package provides functions for cleaning and calibrating low-cost sensor data. It is tailored for two-channel PurpleAir sensors. Functions include those for inspecting and cleaning PurpleAir data, partitioning data for regression for calibration approach, creating multiple calibration models, and evaluating models.
}
\author{
\strong{Maintainer}: Cecilia Martindale \email{cm763@uw.edu} (\href{https://orcid.org/0009-0006-7983-8990}{ORCID})

Authors:
\itemize{
  \item Elena Austin \email{elaustin@uw.edu} (\href{https://orcid.org/0000-0002-4724-1042}{ORCID})
}

}
\keyword{internal}
