% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cal_models.R
\name{model_test_performance}
\alias{model_test_performance}
\title{Predict PM2.5 using selected model and testing data, display performance metrics}
\usage{
model_test_performance(test_data, model)
}
\arguments{
\item{test_data}{test data obtained by partitioning full dataset}

\item{model}{selected calibration model}
}
\value{
plot of observed versus predicted PM2.5
}
\description{
Predict PM2.5 using selected model and testing data, display performance metrics
}
\examples{
test_data <- train_test$test
model_test_performance(test_data, select_model)
}
