% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_clean.R
\name{pa_corr_ab}
\alias{pa_corr_ab}
\title{Get correlation of a and b channels by machine address}
\usage{
pa_corr_ab(
  data,
  channela = NULL,
  channelb = NULL,
  threshold = NULL,
  datetime = NULL
)
}
\arguments{
\item{data}{a dataframe of PurpleAir data, including the mac_address column}

\item{channela}{channel A variable, default is pm2_5_cf_1.}

\item{channelb}{channel B variable, default is pm2_5_cf_1_b.}

\item{threshold}{a PM2.5 value to assess correlation under. Default is 100 ug/m^3.}

\item{datetime}{datetime variable, default is datetime.}
}
\value{
a printed dataframe of correlation between channels, maximum PM2.5,
median PM2.5, mean PM2.5, number of observations, and machine address
(mac_address).
}
\description{
Get correlation of a and b channels by machine address
}
\examples{
pa_corr_ab(data, channela=pm2_5_atm, channelb=pm2_5_atm_b, threshold=500)
pa_corr_ab(data, threshold=150, datetime=UTCDateTime)
}
