% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_clean.R
\name{pa_corr_plot}
\alias{pa_corr_plot}
\title{Title}
\usage{
pa_corr_plot(data, channela = NULL, channelb = NULL)
}
\arguments{
\item{data}{dataframe containing PurpleAir data}

\item{channela}{channel A variable, default is pm2_5_cf_1.}

\item{channelb}{channel B variable, default is pm2_5_cf_1_b.}
}
\value{
plot of correlation of A and B channels.
}
\description{
Title
}
\examples{
pa_corr_plot(pa_data)

pa_corr_plot(pa_data, "pm2_5_atm", "pm2_5_atm_b")
}
