% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_clean.R
\name{pa_timeseries}
\alias{pa_timeseries}
\title{Title}
\usage{
pa_timeseries(data, timescale, channela = NULL, channelb = NULL)
}
\arguments{
\item{data}{dataframe containing PurpleAir data}

\item{timescale}{time period string to show on the x-axis.}

\item{channela}{channel A variable, default is pm2_5_cf_1.}

\item{channelb}{channel B variable, default is pm2_5_cf_1_b.}
}
\value{
timeseries plots of channel A and channel B on the same plot
}
\description{
Title
}
\examples{
pa_timeseries(pa_data, "1 month")
pa_timeseries(pa_data, "3 weeks")
}
