% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa_unclean}
\alias{pa_unclean}
\title{Uncleaned PurpleAir data for examples}
\format{
\subsection{\code{pa_unclean}}{

A data frame with 1,768 rows and 26 columns:
\describe{
\item{mac_address}{Machine address}
\item{current_dewpoint_f, current_temp_f}{Some of the PurpleAir channels}
...
}
}
}
\source{
Martindale et al 2025 (DOI to come)
}
\usage{
pa_unclean
}
\description{
A subset of data from Martindale et al 2025 to be used in examples
}
\keyword{datasets}
