% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cal_models.R
\name{partition}
\alias{partition}
\title{Partition data for calibration using time-sliced approach}
\usage{
partition(data, proportion, initial_Window, seed_val)
}
\arguments{
\item{data}{air quality data for calibration. Should include channels of
interest for calibration and regulatory reference values for matching time
period.}

\item{proportion}{proportion of data to be used in the training dataset (example: 0.7).}

\item{initial_Window}{initial number of observations to be included in a timeslice.
If using hourly data, 168 is one week of hourly observations.}

\item{seed_val}{value to set seed.}
}
\value{
a list of two dataframes, one of training data and one of testing data.
}
\description{
Partition data for calibration using time-sliced approach
}
\examples{
train_test <- partition(cal_data, 0.7, 336, 5)
train_test_two_week_slice <- partition(cal_data, 0.7, 168, 5)
}
