% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_clean.R
\name{prep_pa_data}
\alias{prep_pa_data}
\title{Prepare PurpleAir data for calibration}
\usage{
prep_pa_data(
  data,
  channel = NULL,
  low_threshold = NULL,
  high_threshold = NULL,
  avgtime = NULL,
  station = NULL
)
}
\arguments{
\item{data}{PurpleAir data in a dataframe. This function assumes that the order
of the PurpleAir columns has not been modified. Columns can have been added
to the end, but the default PurpleAir export should not have changed.}

\item{channel}{Channel to filter data based on. The default is pm2_5_cf_ave,
the averaged cf PM2.5 channel. We recommend using an averaged column such
as pm2_5_atm_ave, pm2_5_cf_ave, or pm2.5_aqi_cf_ave.}

\item{low_threshold}{Lower threshold to filter out. Default is 2 ug/m^3 because
this is the limit of detection for (QUESTION: regulatory or PA??) monitors.
We recommend inspecting your data well to determine an appropriate threshold
value for cleaning.}

\item{high_threshold}{Higher threshold to filter out. Default is 500
because it is uncommon to have legitimate values above 500 ug/m^3 outside of
periods of high smoke. We recommend inspecting your data well to determine
an appropriate threshold value for cleaning.}

\item{avgtime}{time period to average measurements over. The default is 60 minutes,
which provides hourly air quality measurements.}

\item{station}{name of station to calibrate off of.}
}
\value{
a dataframe of PurpleAir data ready for calibration, averaged by avgtime
(default is one hour). Data should be thoroughly expected for agreement between
a and b channels and obvious dysfunction before this function.
}
\description{
Prepare PurpleAir data for calibration
}
\examples{
pa_data_cal <- prep_pa_data(pa_data)
pa_data_cal <- prep_pa_data(pa_data, channel="pm2.5_aqi_cf_ave", low_threshold=0,
 high_threshold=300)
}
