% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_clean.R
\name{read_pa}
\alias{read_pa}
\title{Read in PurpleAir files}
\usage{
read_pa(path, timezoneval)
}
\arguments{
\item{path}{the path to the PurpleAir data (folder of csv files).}

\item{timezoneval}{the timezone local to where the PurpleAir data was collected.
To get the list of all timezone names, use OlsonNames().}
}
\value{
dataframe of PurpleAir observations
}
\description{
Read in PurpleAir files
}
\examples{
path <- system.file("extdata/PA_DATA", package = "lcscal")
pa_data <- read_pa(path, "US/Pacific")
}
