% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_data}
\alias{reg_data}
\title{Cleaned reference data}
\format{
\subsection{\code{reg_data}}{

A data frame with 5,214 rows and 5 columns:
\describe{
\item{datetime}{Hourly date-time}
\item{ref_PM25}{Reference PM2.5 (ug/m^3)}
...
}
}
}
\source{
Martindale et al 2025 (DOI to come)
}
\usage{
reg_data
}
\description{
A subset of data from Martindale et al 2025 to be used in examples
}
\keyword{datasets}
