% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{select_model}
\alias{select_model}
\title{Selected model example}
\format{
\subsection{\code{select_model}}{

A list of 25:
\describe{
\item{method}{"lm"}
\item{modelInfo}{Information about the model}
\item{modelType}{Type of model}
...
}
}
}
\source{
Martindale et al 2025 (DOI to come)
}
\usage{
select_model
}
\description{
A subset of data from Martindale et al 2025 to be used in examples.
This is not the selected model from the paper -- it is only to be used in examples.
}
\keyword{datasets}
