% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_test}
\alias{train_test}
\title{Partitioned testing and training data for examples}
\format{
\subsection{\code{train_test}}{

A list of 2:
\describe{
\item{train}{tibble with 1,176 obs. of 26 var}
\item{test}{df with 553 obs. of 24 var}
...
}
}
}
\source{
Martindale et al 2025 (DOI to come)
}
\usage{
train_test
}
\description{
A subset of data from Martindale et al 2025 to be used in examples.
}
\keyword{datasets}
