% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSCal_model}
\alias{TSCal_model}
\title{Time-sliced calibration model}
\format{
\subsection{\code{TSCal_model}}{

A train model with model results, training data, regression model information.
\describe{
\item{results}{Model results}
\item{bestTune}{Training model output of bestTune}
\item{trainingData}{Training data used}
\item{ptype}{Training model output - not used in calibration}
\item{resample}{Results of resampling}
\item{method}{Regression method}
\item{modelInfo}{Model info}
\item{modelType}{Model type}
\item{pred}{Predictions (null) - not used in calibration}
\item{call}{Regression call}
\item{dots}{Empty list - not used in calibration}
\item{metric}{Metric: RMSE}
\item{control}{More resampling data}
\item{finalModel}{Beta-coefficients of final model}
\item{preProcess}{Null - not used in calibration}
\item{resampledCM}{Null - not used in calibration}
\item{perfNames}{Performance metrics - RMSE, R-squared, MAE}
\item{maximize}{Maximize? False}
\item{yLimits}{Limits of y}
\item{times}{Time taken to train model}
\item{levels}{NA - not used in calibration}
\item{terms}{Regression model terms}
\item{coefnames}{Coefficient names}
\item{contrasts}{Contrasts}
\item{xlevels}{Levels of season}
...
}
}
}
\source{
Martindale et al 2026 (DOI to come)
}
\usage{
TSCal_model
}
\description{
A subset of data from Martindale et al 2026 to be used in examples.
This IS the model from the paper.
}
\keyword{datasets}
