% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_PA_data}
\alias{calibrate_PA_data}
\title{Calibrate PA data based on regression model}
\usage{
calibrate_PA_data(data, model)
}
\arguments{
\item{data}{\itemize{
\item to add description
}}

\item{model}{\itemize{
\item to add description
}}
}
\value{
calibrated PA data
}
\description{
Calibrate PA data based on regression model
}
\examples{
to add
}
