% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_PA_data}
\alias{calibrate_PA_data}
\title{Calibrate PA data based on regression model}
\usage{
calibrate_PA_data(data, model)
}
\arguments{
\item{model}{}
}
\description{
Calibrate PA data based on regression model
}
