% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cv.R
\name{do_CV}
\alias{do_CV}
\title{Do cross-validation ADD language from paper}
\usage{
do_CV(data, id, group, formula)
}
\arguments{
\item{data}{dataframe containing the training data.}

\item{id}{unique variable for determining sort order of data frame}

\item{group}{grouping variable (a variable in the data frame)}

\item{formula}{formula to pass to lm}
}
\value{
dataset with a new variable called cvpreds appended to the end;
these are the out-of-sample predictions
}
\description{
Do cross-validation ADD language from paper
}
\examples{
formula <- ref_PM25 ~ 0 + p_0_3_um_ave + p_0_5_um_ave + p_1_0_um_ave +
    p_2_5_um_ave + current_temp_f
train_data <- train_test$train
do_CV(train_data, id = "train_index", group = "select_index", formula)
}
