% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cv.R
\name{get_MSE}
\alias{get_MSE}
\title{Get mean-squared error (MSE)}
\usage{
get_MSE(obs, pred, AIC_var)
}
\arguments{
\item{obs}{outcome variable}

\item{pred}{prediction variable from a model}

\item{AIC_var}{AIC variable}
}
\value{
dataframe of RMSE, MSE-based R^2, and AIC.
}
\description{
Get mean-squared error (MSE)
}
\examples{
one_MSE <- get_MSE(train_data_CV$NPH_PM25, train_data_CV$cvpreds_model1, train_data_CV$AIC_model1)
rbind(get_MSE(train_data_CV$NPH_PM25, train_data_CV$cvpreds_model_1, train_data_CV$AIC_model_1),
  get_MSE(train_data_CV$NPH_PM25, train_data_CV$cvpreds_model_2, train_data_CV$AIC_model_2),
  get_MSE(train_data_CV$NPH_PM25, train_data_CV$cvpreds_model_3, train_data_CV$AIC_model_3),
  get_MSE(train_data_CV$NPH_PM25, train_data_CV$cvpreds_model_4, train_data_CV$AIC_model_4))
}
