% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_clean.R
\name{merge_reg_pa}
\alias{merge_reg_pa}
\title{Merge regulatory and PurpleAir data}
\usage{
merge_reg_pa(pa_data, reg_data, add_season, time_var = NULL)
}
\arguments{
\item{pa_data}{dataframe with cleaned PA data.}

\item{reg_data}{dataframe with cleaned regulatory data.}

\item{add_season}{binary 0 or 1, if 1 will create a season variable based on
the time.}

\item{time_var}{name of time variable, if different from the default of datetimehr.}
}
\value{
a dataframe of merged regulatory and PurpleAir data ready for calibration.
}
\description{
Merge regulatory and PurpleAir data
}
\examples{
cal_data <- merge_reg_pa(pa_data, reg_data, add_season=1)
}
