% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_clean.R
\name{pa_corr_plot}
\alias{pa_corr_plot}
\title{Plot correlation of A and B channels}
\usage{
pa_corr_plot(data, channela = NULL, channelb = NULL, threshold = NULL)
}
\arguments{
\item{data}{dataframe containing PurpleAir data.}

\item{channela}{channel A variable, default is pm2_5_cf_1.}

\item{channelb}{channel B variable, default is pm2_5_cf_1_b.}

\item{threshold}{a PM2.5 value to assess correlation under. Default is 500 ug/m^3.}
}
\value{
plot of correlation of A and B channels.
}
\description{
Plot correlation of A and B channels
}
\examples{
pa_corr_plot(pa_unclean)

pa_corr_plot(pa_unclean, "pm2_5_atm", "pm2_5_atm_b")
}
