% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_clean.R
\name{read_reg}
\alias{read_reg}
\title{Read in regulatory data, filter low values, set datetime interval}
\usage{
read_reg(
  path,
  timezoneval,
  timeformat,
  time_var = NULL,
  PM25_var = NULL,
  avgtime = NULL,
  low_threshold = NULL
)
}
\arguments{
\item{path}{path to regulatory csv files. It is probably worth reading in
one csv file using read_csv to get an idea of the structure and time formatting
before loading in a list.}

\item{timezoneval}{the timezone local to where the regulatory data was collected.
To get the list of all timezone names, use OlsonNames().}

\item{timeformat}{time format of the regulatory data (see ?strptime() for options).}

\item{time_var}{string name of time variable in regulatory csv file.}

\item{PM25_var}{string name of PM2.5 variable in regulatory csv file.}

\item{avgtime}{time interval to average measurements over. The default is 60 minutes.}

\item{low_threshold}{Lower threshold to filter out. Default is 2 ug/m^3 as this
is the limit of the detection for many nephelometers.}
}
\value{
a dataframe of regulatory data ready for use in calibration.
}
\description{
Read in regulatory data, filter low values, set datetime interval
}
\examples{
reg_data <- read_reg("DATA/REG", "US/Pacific", "\%m/\%d/\%Y \%I:\%M:\%S \%p",
  skip=8, time_var=`Observation Time`, PM25_var=`Tukwila Allentown - Pm25Neph - ug/m3 - 1Hr Avg`)
}
