% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cal_models.R
\name{train_model_helper}
\alias{train_model_helper}
\title{ADD reference to this one above once we have a reference. Also add more from the paper.
\subsection{Above: add a note saying that this assumes your data is complete. Pull from paper about}{

Helper function to train models. Can also be used to train an individual model.
}}
\usage{
train_model_helper(formula, train_data)
}
\arguments{
\item{formula}{Regression formula for use in calibration.}

\item{train_data}{Training dataframe.}
}
\value{
results of the model
}
\description{
ADD reference to this one above once we have a reference. Also add more from the paper.
\subsection{Above: add a note saying that this assumes your data is complete. Pull from paper about}{

Helper function to train models. Can also be used to train an individual model.
}
}
\examples{
train_data <- train_test$train
individual_model <- train_model_helper(formula=ref_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity +
   I(current_humidity^2)+
   current_temp_f +
   as.factor(season), train_data)
}
