% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cal_models.R
\name{train_model_helper}
\alias{train_model_helper}
\title{Helper function to train models. Can be used to train an individual model.}
\usage{
train_model_helper(formula, train_data)
}
\arguments{
\item{formula}{Regression formula for use in calibration.}

\item{train_data}{Training dataframe.}
}
\value{
results of the model
}
\description{
Helper function to train models. Can be used to train an individual model.
}
\examples{
train_test_data <- partition(cal_data, 0.7, 168, 5)
train_data <- train_test_data$train
individual_model <- train_model_helper(formula=NPH_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity_outdoor +
   I(current_humidity_outdoor^2)+
   current_temp_f_outdoor +
   as.factor(season), train_data)
}
