% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cal_models.R
\name{train_models}
\alias{train_models}
\title{Train multiple models and compare results.}
\usage{
train_models(formulas, train_data)
}
\arguments{
\item{formulas}{a named list of formulas. ADD MORE FROM PAPER}

\item{train_data}{training dataframe. ADD MORE FROM PAPER}
}
\value{
summaries of models being trained and resampled plot of RMSE.
}
\description{
Train multiple models and compare results.
}
\examples{
train_test_data <- partition(cal_data, 0.7, 168, 5)
train_data <- train_test_data$train
formulas <- list(formula1 = NPH_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity + I(current_humidity^2) +
   current_temp_f + as.factor(season),
                 formula2 = NPH_PM25 ~ 0 + pm2_5_cf_ave +
   current_humidity + current_temp_f + as.factor(season),
                 formula3 = NPH_PM25 ~ 0 + pm2_5_cf_ave +
   current_temp_f + as.factor(season),
                 formula4 = NPH_PM25 ~ 0 + pm2_5_cf_ave + as.factor(season))
train_models(formulas, train_data)
}
